+++ CHANGELOG.md
@@ -17,3 +17,10 @@
 
 ## 1.2.2
 - A few updates for Nomx, basically more properties. And a new class, RippleButton.
+
+## 1.3.0
+- Fixes to typescript compiler that *should* work with folders with spaces. In the future I do want to have the folder contain node_modules for type definitions for libraries, so that
+	```ts
+	import Library from "@cdo-sync/library"
+	```
+	Can be used instead of adding to add each manually and have all the library type definitions pre-defined.